+++ src/main/java/school/sptech/back_end_PI/repository/ContratoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import school.sptech.back_end_PI.entity.Aula;
 import school.sptech.back_end_PI.entity.Contrato;
 
-public interface ContratoRepository extends JpaRepository<Contrato, Integer>  {
+public interface ContratoRepository extends JpaRepository<Contrato, Long>  {
 }
